refactor(installation): split monolithic module into structured submodules
