+++ src-tauri/src/gui/commands/mod.rs
@@ -1,6 +1,6 @@
-pub mod utils_commands;
-pub mod prequisites;
-pub mod installation;
-pub mod settings;
 pub mod idf_tools;
+pub mod installation;
+pub mod prequisites;
+pub mod settings;
+pub mod utils_commands;
 pub mod version_management;
